requirements.txt
discord
requests
BeautifulSoup4
load_dotenv
selenium
pymongo
reportlab